requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
PyJWT==2.8.0